DSS-3365: Rules for loading information 1. gateway loading plan get api done type: feature

+++ api/cp-dss-api/loading-plan/src/main/java/com/cpdss/loadingplan/entity/LoadingRule.java
@@ -36,10 +36,10 @@ public class LoadingRule extends EntityDoc {
   private Long loadingXid;
 
   @Column(name = "numeric_precision")
-  private Integer numericPrecision;
+  private Long numericPrecision;
 
   @Column(name = "numeric_scale")
-  private Integer numericScale;
+  private Long numericScale;
 
   @Column(name = "parent_rule_xid")
   private Long parentRuleXid;
